Random loading view and tag convertion.
Randomly pick one loading view style to use in each time.
Refactor the id->tag convertion to String extension.
