+++ init.engineer-iOS-App/Controller/ReviewTabController.swift
@@ -20,7 +20,7 @@ class ReviewTabController: UIViewController {
     var adBanner = GADBannerView(adSize: kGADAdSizeMediumRectangle)
     var interstitial = GADInterstitial(adUnitID: K.getInfoPlistByKey("GAD AdsInterstitial") ?? "")
     
-    var loadingView = NVActivityIndicatorView(frame: CGRect(x: 0, y: 0, width: 50, height: 50), type: .audioEqualizer, color: .cyan, padding: .none)
+    var loadingView = NVActivityIndicatorView(frame: CGRect(x: 0, y: 0, width: 50, height: 50), type: .randomPick(), color: .cyan, padding: .none)
     let GAP_ID = "gap"
     let REVIEW_ID = "review"
     let TITLE_ID = "title"
